Allow for measuring the source up-to-dateness of Trivy JSON reports.
Closes #10608.
