+++ components/collector/.vulture_ignore_list.py
@@ -148,18 +148,19 @@ SonarQubeUncoveredBranches  # unused class (src/source_collectors/sonarqube/unco
 SonarQubeUncoveredLines  # unused class (src/source_collectors/sonarqube/uncovered_lines.py:6)
 TrelloIssues  # unused class (src/source_collectors/trello/issues.py:13)
 TrelloSourceUpToDateness  # unused class (src/source_collectors/trello/source_up_to_dateness.py:12)
-VulnerabilityID  # unused variable (src/source_collectors/trivy/security_warnings.py:19)
-Title  # unused variable (src/source_collectors/trivy/security_warnings.py:20)
-Description  # unused variable (src/source_collectors/trivy/security_warnings.py:21)
-PkgName  # unused variable (src/source_collectors/trivy/security_warnings.py:23)
-InstalledVersion  # unused variable (src/source_collectors/trivy/security_warnings.py:24)
-FixedVersion  # unused variable (src/source_collectors/trivy/security_warnings.py:25)
-References  # unused variable (src/source_collectors/trivy/security_warnings.py:26)
-Target  # unused variable (src/source_collectors/trivy/security_warnings.py:32)
-Vulnerabilities  # unused variable (src/source_collectors/trivy/security_warnings.py:33)
-SchemaVersion  # unused variable (src/source_collectors/trivy/security_warnings.py:49)
-Results  # unused variable (src/source_collectors/trivy/security_warnings.py:50)
-TrivyJSONSecurityWarnings  # unused class (src/source_collectors/trivy/security_warnings.py:56)
+VulnerabilityID  # unused variable (src/source_collectors/trivy/base.py:15)
+Title  # unused variable (src/source_collectors/trivy/base.py:16)
+Description  # unused variable (src/source_collectors/trivy/base.py:17)
+PkgName  # unused variable (src/source_collectors/trivy/base.py:19)
+InstalledVersion  # unused variable (src/source_collectors/trivy/base.py:20)
+FixedVersion  # unused variable (src/source_collectors/trivy/base.py:21)
+References  # unused variable (src/source_collectors/trivy/base.py:22)
+Target  # unused variable (src/source_collectors/trivy/base.py:28)
+Vulnerabilities  # unused variable (src/source_collectors/trivy/base.py:29)
+SchemaVersion  # unused variable (src/source_collectors/trivy/base.py:45)
+Results  # unused variable (src/source_collectors/trivy/base.py:46)
+TrivyJSONSecurityWarnings  # unused class (src/source_collectors/trivy/security_warnings.py:12)
+TrivyJSONSourceUpToDateness  # unused class (src/source_collectors/trivy/source_up_to_dateness.py:10)
 VisualStudioTRXSourceUpToDateness  # unused class (src/source_collectors/visual_studio_trx/source_up_to_dateness.py:13)
 totalCount  # unused variable (tests/source_collectors/github/test_merge_requests.py:16)
 baseRefName  # unused variable (tests/source_collectors/github/test_merge_requests.py:24)
@@ -173,4 +174,3 @@ endCursor  # unused variable (tests/source_collectors/github/test_merge_requests
 pageInfo  # unused variable (tests/source_collectors/github/test_merge_requests.py:45)
 totalCount  # unused variable (tests/source_collectors/github/test_merge_requests.py:46)
 pullRequests  # unused variable (tests/source_collectors/github/test_merge_requests.py:53)
-RobotFrameworkSourceVersion  # unused class (tests/source_collectors/robot_framework/test_source_version.py:6)
